se logra conectar el json enviado con el template
