+++ package.json
@@ -14,7 +14,8 @@
         "body-parser": "^1.19.2",
         "dotenv": "^16.0.0",
         "express": "^4.17.3",
-        "hbs": "^4.2.0"
+        "hbs": "^4.2.0",
+        "jquery": "^3.6.0"
     },
     "devDependencies": {
         "eslint": "^7.25.0",
