+++ public/css/style.css
@@ -13,76 +13,124 @@ body{
 
 .main-content{
     flex-grow: 1;
+    padding-top: 3%;
 }
 
 h1{
     color: rgb(4, 252, 252);
     font-size: 45px;
+    margin-bottom: 0px;
 }
 
-span, label, #message-1{
+span, #message-1, #welcome{
     color: cyan;
-    font-size: 1.0em;
+    margin: 2% 1% 1% 1%;
+}
+
+label{
+    color: cyan;
+    font-size: smaller;
+    margin: 2% 1% 1% 1%;
 }
 #bodycontent{
     padding: 0% 2%;
 }
-#message-1, #message-2{
-    font-size: 1.2em;
-}
+
 img{
     height: 43vh;
     margin: 0px;
     padding: 0px;
 }
-a, button, .button{
+button, .button{
     text-decoration: none;
     color: rgb(54, 5, 168);
     background-color: aqua;
     padding: 4px;
-    margin: 3px;
-    border-radius: 5%;
+    margin: 1%;
+    border-radius: 3px;
     border: none;
-    
-
 }
-button{
+.button{
     cursor:grab;
+}
+a{
+    text-decoration: none;
+    color: rgb(54, 5, 168);
+    background-color: aqua;
+    padding: 4px 3px;
+    margin: 5% 2% 7% 0%;
+    border-radius: 3px;
+    border: none;
+    line-height: 2em;
 }
 a:hover{
     background-color: rgb(32, 3, 112);
-    color: rgb(255, 0, 183);
+    color: cyan;
 }
 
-button:hover{
-    background-color: rgb(105, 250, 2);
-    color: rgb(76, 0, 255);
+.button:hover{
+    background-color: rgb(32, 3, 112);
+    color: cyan;
 }
 #footer{
     border-top: 1px dotted rgb(2, 70, 172);
-    margin-top: 12px;
-    margin-bottom: 12px;
-    padding: 14px;
+    margin-top: 10px;
+    margin-bottom: 10px;
+    padding: 10px;
     background-color: rgb(6, 108, 175);
     color:rgb(23, 4, 190)
 }
 #header{
-    margin: 12px 0px;
-    padding: 14px;
+    margin: 0;
+    padding: 5px 10px;
     background-color: rgb(6, 108, 175);
     
 }
 
 .cajaresultados{
-    margin-top: 2%;
-    padding: 2%;
+    margin-top: 0%;
+    padding: 1%;
+    overflow-y: scroll; 
+    height: 35vh; 
+    border-radius: 3px; background-color: rgb(229, 247, 241);
 }
 
 #formulario{
     margin-top: 10px;
 }
 
-#message-2{
-    color:blue;
+
+table {
+    table-layout: fixed;
+    width: 100%;
+    border-collapse: collapse;
+}
+th{
+    color:rgb(7, 7, 172); 
+    background-color: rgb(4, 252, 252);
+}
+td{
+    color:rgb(7, 7, 143);
     font-size: smaller;
+}
+
+thead th:nth-child(1) {
+    width: 30%;
+}
+
+thead th:nth-child(2) {
+    width: 20%;
+}
+
+thead th:nth-child(3) {
+    width: 15%;
+}
+
+thead th:nth-child(4) {
+    width: 35%;
+}
+
+th, td {
+    padding: 5px;
+    text-align: left;
 }
