--- templates/searchinrange.hbs
@@ -1,50 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="icon" href="/img/nubi.jpg">
-    <link rel="stylesheet" href="css/style.css">
-    <title>searchinrange</title>
-</head>
-<body>
-    <div id="header">
-        <h1>{{title}}</h1>
-    </div>
-    <div class="main-content">
-        <div id="bodycontent">
-            <p>Use this site to find out birthdays in Nisum!!</p>
-            <div>
-                <a href="/">Home</a>
-                <a href="/thismonth" class="buttonTM" >This month</a>
-                <a href="/nextmonth" class="buttonNM" >Next Month</a>
-            </div>
-            
-            <form id='formulario' action="" style="padding: 2px 2px; border-radius: 3px; border: none;">
-                <div>
-                    <label for="birthday">Start Date</label>
-                    <input type="date" id="Startdate" name="startDate" class="form-control"
-                    value="2022-01-01"
-                    min="2022-01-01" max="2022-12-31">
-                    <label for="birthday">End Date</label>
-                    <input type="date" id="Enddate" name="endDate" class="form-control"
-                    value="2022-07-22"
-                    min="2022-01-01" max="2022-12-31">
-                    <input class="button" type="submit" value="Search!!">
-                </div>
-            </form> 
-                <p id="message-1"></p>
-            <div class="cajaresultados border border-muted" style="overflow-y: scroll; height: 25vh; border-radius: 3px; background-color: aqua;">  
-                <p id="message-2"></p>
-            </div>
-        </div>
-    </div>
-    
-    <div id="footer">
-        <span>Created by FL</span>
-    </div id="footer">
-
-    <script src="js/app.js"></script>
-</body>
-</html>
