--- templates/thismonth.hbs
@@ -1,40 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="icon" href="/img/nubi.jpg">
-    <link rel="stylesheet" href="css/style.css">
-    <title>{{title}}</title>
-</head>
-<body>
-    <div id="header">
-        <h1>{{title}}</h1>
-    </div>
-    <div class="main-content">
-        <div id="bodycontent2">
-            <p>Use this site to find out birthdays in Nisum!!</p>
-            <div>
-                <a href="/searchinrange">Search in a Range</a>
-                <a href="/">Home</a>
-                <a  href="/nextmonth" class="buttonNM">Next month</a>
-                
-            </div>
-                <p id="message-3"></p>
-            <div class="cajaresultados border border-muted" style="overflow-y: scroll; height: 25vh; border-radius: 3px; background-color: aqua;">  
-                <p style="color: blue;" id="message-4"></p>
-            </div>
-        </div>
-    </div>
-    
-    <div id="footer">
-        <span>Created by FL</span>
-    </div id="footer">
-    <!--<script src="js/jquery-1.11.0.min.js" type="text/javascript"></script>
-    <script type="text/javascript">
-    
-    </script>*/-->
-
-    <script src="js/app.js"></script>
-</body>
